perf(decode): narrow mi-size grid element type from usize to u8 (#1067)
* perf(decode): narrow mi-size grid element type from usize to u8

Feature ID: INFRA-DECODE-SERIAL-HOT-PATHS

Allocation campaign slice 3 (traffic). The coalesced TileMiSizeState backing
store holds block-size indices (< 29) and partition-context values (<= 63),
which fit in a byte but were stored as usize -- 8x wider than needed. Narrow
the storage to Vec<u8>. The partition-decision read path
(TilePartitionContextState, PartitionContextInput, SquareSplitContextInput,
and the grid/neighbour leaf helpers in cdf/context) now carries &[u8] and
widens to usize at the point of use; the block-size index is converted with a
checked u8::try_from that surfaces a typed error, preserving the no-panic
discipline. Values round-trip identically, so decode output is unchanged.

Measured (single-thread, in-container; no dav2d/1080p test.ivf available):
- DHAT allocation traffic: TIP 5,777,818 -> 5,697,178 bytes (-80.6 KB,
  -1.4%); opfl 3,682,490 -> 3,650,234 (-32.3 KB, -0.88%). Block count
  unchanged (same allocations, narrower). The saving scales with tile mi
  area, so it is larger at real resolutions than on these micro-fixtures.
- callgrind Ir: TIP -0.04%, opfl -0.05% (u8 loads/stores are cheaper).
- Byte-exact on all five fixtures; splot-decode suite green (1,643 tests).


* fix(decode): scale mi-size byte limit by u8 element size

Feature ID: INFRA-DECODE-SERIAL-HOT-PATHS

Follow-up to the mi-size grid narrowing: the backing store is now Vec<u8>,
but ensure_mi_size_allocation_within_limits still multiplied the element
count by size_of::<usize>(), so the MaxDecodedFrameBytes guard over-counted
this allocation 8x and could reject a tile frontier that the narrowed store
would fit. Multiply by size_of::<u8>() to match the real element width. The
two byte-budget tests are updated to the true 576-byte size and a threshold
below it so they still exercise the guard.


* test(decode): merge duplicate mi-size byte-budget frontier tests

Feature ID: INFRA-DECODE-SERIAL-HOT-PATHS

The two mi-size byte-budget tests became identical after the u8 limit fix
(differing only in which frontier function they call), tripping the
incremental duplication gate. Extract the shared setup/assert into
assert_mi_state_byte_budget_rejected and drive both the partition and
LR-unit frontier paths from a single test, so neither frontier's guard
coverage is lost and no duplicate function remains.


---------

+++ crates/splot-decode/src/bitstream/tile_payload/general_intra_block_tests.rs
@@ -23,12 +23,12 @@ const BLOCK_16X16: usize = 6;
 const BLOCK_8X8: usize = 3;
 const BLOCK_32X16: usize = 8;
 const BLOCK_64X64: usize = 12;
-const BLOCK_256X256: usize = 18;
+const BLOCK_256X256: u8 = 18;
 const BLOCK_4X16: usize = 19;
 const TX_4X4: usize = 0;
 const TX_16X16: usize = 2;
 const TX_32X32: usize = 3;
-const CLEAR_PARTITION_CONTEXT: usize = 0;
+const CLEAR_PARTITION_CONTEXT: u8 = 0;
 const PAYLOAD: [u8; 2] = [0x12, 0xFB];
 
 fn make_work_unit(payload: &[u8]) -> DecodeTileWorkUnit<'_> {
